moved exception to phpdocumentor namepasce
