+++ src/phpDocumentor/Project/Version/DefinitionFactory.php
@@ -13,7 +13,7 @@ namespace phpDocumentor\Project\Version;
 
 use phpDocumentor\DocumentGroupDefinitionFactory;
 use phpDocumentor\DocumentGroupFormat;
-use phpDocumentor\Exception\Exception;
+use phpDocumentor\Exception;
 use phpDocumentor\Project\DocumentGroup\Definition as DocumentGroupDefinition;
 use phpDocumentor\Project\VersionNumber;
